ci: drop pytest in favor of Sprocket test (#263)
~~This PR is for demonstration purposes of the RFC document
[here](https://stjude-rust-labs.github.io/rfcs/branches/rfcs/sprocket-test/0001-sprocket-test.html)~~

this PR has morphed from a demonstration into a real PR for swapping our
test infra from pytest to Sprocket. Still a WIP as not all the tests are
written yet and it has not been hooked into CI yet

---------

+++ tools/picard.wdl
@@ -139,7 +139,10 @@ task validate_bam {
         description: "Validates the input BAM file for correct formatting using Picard"
         external_help: "https://gatk.broadinstitute.org/hc/en-us/articles/360057440611-ValidateSamFile-Picard-"
         outputs: {
-            validate_report: "Validation report produced by `picard ValidateSamFile`. Validation warnings and errors are logged.",
+            validate_report: {
+                description: "Validation report produced by `picard ValidateSamFile`. Validation warnings and errors are logged.",
+                help: "If `summary_mode` is not enabled, then this file will be gzipped and contain one line for every warning or error logged by Picard. If `summary_mode == true` this will be a plaintext summary of warnings and errors encountered by Picard.",
+            },
         }
     }
 
@@ -161,19 +164,18 @@ task validate_bam {
             ],
             tool_default: "STRICT",
         }
-        succeed_on_errors: {
-            description: "Succeed the task even if errors *and/or* warnings are detected",
+        deny_errors: {
+            description: "Fail the task if any errors are detected by Picard.",
             group: "Common",
         }
-        succeed_on_warnings: {
-            description: "Succeed the task if warnings are detected and there are no errors. Overridden by `succeed_on_errors`",
+        deny_warnings: {
+            description: "Fail the task if any errors or warnings are detected by Picard.",
             group: "Common",
         }
         summary_mode: {
             description: "Enable SUMMARY mode?",
             group: "Common",
         }
-        index_validation_stringency_less_exhaustive: "Set `INDEX_VALIDATION_STRINGENCY=LESS_EXHAUSTIVE`?"
         max_errors: "Set the value of MAX_OUTPUT for `picard ValidateSamFile`. The Picard default is 100, a lower number can enable fast fail behavior"
         memory_gb: "RAM to allocate for task, specified in GB"
         modify_disk_size_gb: "Add to or subtract from dynamic disk space allocation. Default disk size is determined by the size of the inputs. Specified in GB."
@@ -185,10 +187,9 @@ task validate_bam {
         Array[String] ignore_list = []
         String outfile_name = basename(bam, ".bam") + ".ValidateSamFile.txt"
         String validation_stringency = "LENIENT"
-        Boolean succeed_on_errors = false
-        Boolean succeed_on_warnings = true
+        Boolean deny_errors = true
+        Boolean deny_warnings = false
         Boolean summary_mode = false
-        Boolean index_validation_stringency_less_exhaustive = false
         Int max_errors = 2147483647  # max 32-bit INT
         Int memory_gb = 16
         Int modify_disk_size_gb = 0
@@ -199,9 +200,6 @@ task validate_bam {
         else outfile_name + ".gz"
     String mode_arg = if summary_mode
         then "--MODE SUMMARY"
-        else ""
-    String stringency_arg = if index_validation_stringency_less_exhaustive
-        then "--INDEX_VALIDATION_STRINGENCY LESS_EXHAUSTIVE"
         else ""
     Float bam_size = size(bam, "GB")
     Int disk_size_gb = ceil(bam_size) + 10 + modify_disk_size_gb
@@ -216,7 +214,6 @@ task validate_bam {
             -I "~{bam}" \
             ~{"-R '" + reference_fasta + "'"} \
             ~{mode_arg} \
-            ~{stringency_arg} \
             --VALIDATION_STRINGENCY "~{validation_stringency}" \
             ~{sep(" ", prefix("--IGNORE ", squote(ignore_list)))} \
             --MAX_OUTPUT ~{max_errors} \
@@ -232,20 +229,11 @@ task validate_bam {
         # rc = 2 = validation errors and warnings
         # rc = 3 = validation errors (no warnings)
         if [ $rc -ne 0 ] && [ $rc -ne 1 ] && [ $rc -ne 2 ] && [ $rc -ne 3 ]; then
+            >&2 echo "Unexpected exit code from Picard ValidateSamFile"
             exit $rc
-        fi
-
-        if ~{succeed_on_warnings}; then
-            GREP_PATTERN="ERROR"
-        else
-            GREP_PATTERN="(ERROR|WARNING)"
-        fi
-
-        if ! ~{succeed_on_errors} \
-            && [ "$(grep -Ec "$GREP_PATTERN" "~{outfile}")" -gt 0 ]
+        elif [ "$rc" -gt 0 ] && ~{deny_warnings} || [ "$rc" -gt 1 ] && ~{deny_errors}
         then
             >&2 echo "Problems detected by Picard ValidateSamFile"
-            >&2 grep -E "$GREP_PATTERN" "~{outfile}"
             exit $rc
         fi
     >>>
@@ -412,16 +400,23 @@ task merge_sam_files {
             --ASSUME_SORTED true \
             --SORT_ORDER "~{sort_order}" \
             --USE_THREADING ~{threading} \
-            --CREATE_INDEX true \
+            ~{if sort_order == "coordinate"
+                then "--CREATE_INDEX true"
+                else ""
+            } \
             --CREATE_MD5_FILE true \
             --VALIDATION_STRINGENCY "~{validation_stringency}"
 
-        mv "~{prefix}.bai" "~{outfile_name}.bai"
+        # CREATE_INDEX true only applies when the sort order
+        # is coordinate. So the index may not exist.
+        if [ -f "~{prefix}.bai" ]; then
+            mv "~{prefix}.bai" "~{outfile_name}.bai"
+        fi
     >>>
 
     output {
         File merged_bam = outfile_name
-        File merged_bam_index = outfile_name + ".bai"
+        File? merged_bam_index = outfile_name + ".bai"
         File merged_bam_md5 = outfile_name + ".md5"
     }
 
@@ -824,69 +819,6 @@ task quality_score_distribution {
     }
 }
 
-#@ except: MatchingOutputMeta
-task bam_to_fastq {
-    meta {
-        description: "This WDL task converts the input BAM file into FASTQ format files."
-        warning: "**[Deprecated]** This task has been deprecated in favor of `samtools.bam_to_fastq` which is more performant and doesn't error on 'illegal mate states'."
-        deprecated: true
-    }
-
-    parameter_meta {
-        bam: "Input BAM format file to convert to FASTQ"
-        prefix: "Prefix for the <type of file> file. The extension `<extension>` will be added."
-        paired: {
-            description: "Is the data Paired-End (true) or Single-End (false)?",
-            group: "Common",
-        }
-        memory_gb: "RAM to allocate for task, specified in GB"
-        modify_disk_size_gb: "Add to or subtract from dynamic disk space allocation. Default disk size is determined by the size of the inputs. Specified in GB."
-    }
-
-    input {
-        File bam
-        String prefix = basename(bam, ".bam")
-        Boolean paired = true
-        Int memory_gb = 56
-        Int modify_disk_size_gb = 0
-    }
-
-    Float bam_size = size(bam, "GB")
-    Int disk_size_gb = ceil(bam_size * 4) + 10 + modify_disk_size_gb
-    Int java_heap_size = ceil(memory_gb * 0.9)
-
-    command <<<
-        set -euo pipefail
-
-        picard -Xmx~{java_heap_size}g SamToFastq INPUT="~{bam}" \
-            FASTQ="~{prefix}.R1.fastq" \
-            ~{if paired
-                then "SECOND_END_FASTQ='" + prefix + ".R2.fastq'"
-                else ""
-            } \
-            RE_REVERSE=true \
-            VALIDATION_STRINGENCY=SILENT
-
-        gzip "~{prefix}.R1.fastq" \
-            ~{if paired
-                then "'" + prefix + ".R2.fastq'"
-                else ""
-            }
-    >>>
-
-    output {
-        File read_one_fastq_gz = "~{prefix}.R1.fastq.gz"
-        File? read_two_fastq_gz = "~{prefix}.R2.fastq.gz"
-    }
-
-    runtime {
-        memory: "~{memory_gb} GB"
-        disks: "~{disk_size_gb} GB"
-        container: "quay.io/biocontainers/picard:3.1.1--hdfd78af_0"
-        maxRetries: 1
-    }
-}
-
 task merge_vcfs {
     meta {
         description: "Merges the input VCF files into a single VCF file"
@@ -1055,6 +987,69 @@ task create_sequence_dictionary {
 
     runtime {
         cpu: 1
+        memory: "~{memory_gb} GB"
+        disks: "~{disk_size_gb} GB"
+        container: "quay.io/biocontainers/picard:3.1.1--hdfd78af_0"
+        maxRetries: 1
+    }
+}
+
+#@ except: MatchingOutputMeta
+task bam_to_fastq {
+    meta {
+        description: "This WDL task converts the input BAM file into FASTQ format files."
+        warning: "**[Deprecated]** This task has been deprecated in favor of `samtools.bam_to_fastq` which is more performant and doesn't error on 'illegal mate states'."
+        deprecated: true
+    }
+
+    parameter_meta {
+        bam: "Input BAM format file to convert to FASTQ"
+        prefix: "Prefix for the <type of file> file. The extension `<extension>` will be added."
+        paired: {
+            description: "Is the data Paired-End (true) or Single-End (false)?",
+            group: "Common",
+        }
+        memory_gb: "RAM to allocate for task, specified in GB"
+        modify_disk_size_gb: "Add to or subtract from dynamic disk space allocation. Default disk size is determined by the size of the inputs. Specified in GB."
+    }
+
+    input {
+        File bam
+        String prefix = basename(bam, ".bam")
+        Boolean paired = true
+        Int memory_gb = 56
+        Int modify_disk_size_gb = 0
+    }
+
+    Float bam_size = size(bam, "GB")
+    Int disk_size_gb = ceil(bam_size * 4) + 10 + modify_disk_size_gb
+    Int java_heap_size = ceil(memory_gb * 0.9)
+
+    command <<<
+        set -euo pipefail
+
+        picard -Xmx~{java_heap_size}g SamToFastq INPUT="~{bam}" \
+            FASTQ="~{prefix}.R1.fastq" \
+            ~{if paired
+                then "SECOND_END_FASTQ='" + prefix + ".R2.fastq'"
+                else ""
+            } \
+            RE_REVERSE=true \
+            VALIDATION_STRINGENCY=SILENT
+
+        gzip "~{prefix}.R1.fastq" \
+            ~{if paired
+                then "'" + prefix + ".R2.fastq'"
+                else ""
+            }
+    >>>
+
+    output {
+        File read_one_fastq_gz = "~{prefix}.R1.fastq.gz"
+        File? read_two_fastq_gz = "~{prefix}.R2.fastq.gz"
+    }
+
+    runtime {
         memory: "~{memory_gb} GB"
         disks: "~{disk_size_gb} GB"
         container: "quay.io/biocontainers/picard:3.1.1--hdfd78af_0"
